deepLinkAppWidgetProvider mvp
